Use dmix for audio system
Doing this instead of direct hardware access by default because it's is the most compatible method for most situations. The chip runs at 44.1 kHz, with all input being downsampled where necessary and mixed. This allows multiple tasks to access the audio system simultaneously with the tradeoff that they get the same output sample rate.

However, this can be overridden with an ~/.asoundrc file specifying direct hardware access. The first task then gets exclusive access to the STAC9202 and can alter the sample rate arbitrarily.

This feels like a good compromise that won't bother the majority, but enables direct control for those who need it without modifying system files.