Enable auto-muting and admit only ext2 works
There's no really benefit to ext3 or ext4 on flash storage and there's a weird problem where the UUID of an ext4 volume cannot be changed without a fresh e2fsck scan, which you can't run when the volume is mounted, which is a problem for us changing it on the root fs. But ext2 is fine.